+++ pkg/validator/validator.go
@@ -2,7 +2,7 @@ package validator
 
 import (
 	"github.com/go-playground/validator/v10"
-	"github.com/rocketblend/rocketblend/pkg/driver/types"
+	"github.com/rocketblend/rocketblend/pkg/types"
 )
 
 type (
